split up profile component
